WIP adding transform getter method
